chore: code modulairy and refractoring is ensured
